# foundry_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcadAssist API",
    "description": "Backend API powering the AcadAssist AI Study Assistant via Microsoft Foundry.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/tools/generate_quiz": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Generate an practice quiz",
        "description": "Generate practice quiz questions on a specified academic subject or topic.",
        "operationId": "generate_quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateQuizParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate Quiz"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/submit_quiz": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Submit student quiz answers",
        "description": "Submit answers for a previously generated quiz, calculate score, and record history.",
        "operationId": "submit_quiz",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Submit Quiz"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_performance": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve historical student performance metrics",
        "description": "Retrieve historical assessment scores, quiz statistics, and performance trends.",
        "operationId": "get_performance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetPerformanceParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Performance"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools/get_weak_topics": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Identify student weak topics",
        "description": "Identify topics where the student has struggled based on quiz and assessment history.",
        "operationId": "get_weak_topics",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetWeakTopicsParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Weak Topics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tools/get_upcoming_exams": {
      "post": {
        "tags": [
          "Tools"
        ],
        "summary": "Retrieve upcoming exams and deadlines",
        "description": "Retrieve scheduled exams, tests, and assignment deadlines.",
        "operationId": "get_upcoming_exams",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUpcomingExamsParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Upcoming Exams"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerateQuizParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "topic_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {}
            ],
            "title": "Topic Ids"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "description": "Quiz difficulty (easy, medium, hard)"
          },
          "count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Count",
            "description": "Number of questions",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "subject_id",
          "difficulty"
        ],
        "title": "GenerateQuizParams"
      },
...
      "GetPerformanceParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetPerformanceParams"
      },
...
      "GetUpcomingExamsParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetUpcomingExamsParams"
      },
      "GetWeakTopicsParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "subject_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "GetWeakTopicsParams"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubmitQuizParams": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "string",
            "title": "Quiz Id"
          },
          "answers": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {}
            ],
            "title": "Answers",
            "description": "Student's chosen answers"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "quiz_id",
          "answers"
        ],
        "title": "SubmitQuizParams"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://lqs5ss0s-8000.inc1.devtunnels.ms"
    }
  ]
}